chatInfo 컴포넌트 생성 및 chatpage 레이아웃 조정
